+++ indico/modules/core/forms.py
@@ -10,7 +10,6 @@ from wtforms.validators import DataRequired
 
 from indico.util.i18n import _
 from indico.web.forms.base import IndicoForm
-from indico.web.forms.validators import IndicoRegexp
 from indico.web.forms.widgets import SwitchWidget
 
 
@@ -21,8 +20,8 @@ class SettingsForm(IndicoForm):
                                          description=_('The organization that runs this Indico site.'))
 
     # Social settings
-    social_enabled = BooleanField(_('Enabled'), widget=SwitchWidget())
-    social_facebook_app_id = StringField('Facebook App ID', [IndicoRegexp(r'^\d*$')])
+    social_enabled = BooleanField(_('Enabled'), widget=SwitchWidget(),
+                                  description=_('Show a share widget in the footer of event pages.'))
 
     @property
     def _fieldsets(self):